[Presentation] Added wish result item bottom img
