+++ lib/domain/extensions/weapon_type_extensions.dart
@@ -5,4 +5,8 @@ extension WeaponTypeExtension on WeaponType {
   String getWeaponAssetPath() {
     return Assets.getWeaponTypePath(this);
   }
+
+  String getWeaponNormalSkillAssetPath() {
+    return Assets.getWeaponSkillAssetPath(this);
+  }
 }
